requirements.txt
awsebcli==3.15.2
blessed==1.15.0
boto3==1.9.118
botocore==1.12.118
cached-property==1.5.1
cachetools==3.1.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-cors-headers==2.5.2
django-extensions==2.1.6
django-jet==1.0.8
djangorestframework==3.9.2
docker==3.7.1
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
drf-yasg==1.14.0
dynamodb-json==1.3
ecdsa==0.13
envs==1.3
future==0.16.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.13.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
MarkupSafe==1.1.1
oauthlib==3.0.2
pathspec==0.5.9
Pillow==6.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
pypiwin32==223
python-dateutil==2.8.0
pytz==2018.9
pywin32==227
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.2.0
rsa==4.0
ruamel.yaml==0.15.89
s3transfer==0.2.0
semantic-version==2.5.0
sentry-sdk==0.13.5
simplejson==3.16.0
six==1.11.0
stripe==2.41.1
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.55.0
